requirements.txt
flask
flask-sqlalchemy
pyodbc
transformers
torch
scikit-learn
python-dotenv
numpy